requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
crewai[google-genai]>=1.9.3
crewai-tools
python-dotenv>=1.0.0
pydantic>=2.0
httpx>=0.24.0